Calculadora.java
Esta tiene Doubles en lugar de ints

+++ Calculadora/Calculadora.java
@@ -1,19 +1,19 @@
 public class Calculadora{
     public int x,y;
 
-    public static int suma(int x, int y){
+    public static double suma(int x, int y){
         return x+y;
 	}
 
-	public static int resta(int x, int y){
+	public static double resta(int x, int y){
         return  x - y;
     }
     
-    public static int mul(int x, int y){
+    public static double mul(int x, int y){
         return x * y;
     }
 
-    public static int div(int x, int y){
+    public static double div(int x, int y){
         return x / y;
     }
 }
